pyproject.toml
[tool.poetry]
name = "salam bot"
version = "0.1.0"
description = "Telegram VPN WebApp Bot on FastAPI and Postgres"
authors = ["CarloDiPalma <ping.06@mail.ru>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aiosqlite = "0.20.0"
alembic = "1.13.1"
annotated-types = "0.6.0"
anyio = "4.3.0"
bcrypt = "4.1.2"
build = "1.2.1"
cachecontrol = "0.14.0"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
click = "8.1.7"
colorama = "0.4.6"
crashtest = "0.4.1"
cryptography = "42.0.5"
distlib = "0.3.8"
dnspython = "2.6.1"
dulwich = "0.21.7"
email-validator = "2.1.1"
exceptiongroup = "1.2.0"
fastapi = "0.110.0"
fastapi-users = "12.1.3"
fastapi-users-db-sqlalchemy = "6.0.1"
fastjsonschema = "2.20.0"
filelock = "3.15.4"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.4"
httptools = "0.6.1"
httpx = "0.27.0"
idna = "3.6"
importlib-metadata = "8.0.0"
installer = "0.7.0"
isort = "5.13.2"
itsdangerous = "2.1.2"
jaraco-classes = "3.4.0"
jinja2 = "3.1.3"
keyring = "24.3.1"
makefun = "1.15.2"
mako = "1.3.5"
markupsafe = "2.1.5"
more-itertools = "10.3.0"
msgpack = "1.0.8"
orjson = "3.9.15"
packaging = "24.1"
passlib = "1.7.4"
pexpect = "4.9.0"
pkginfo = "1.11.1"
platformdirs = "4.2.2"
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.8.0"
psycopg2 = "2.9.9"
ptyprocess = "0.7.0"
pycparser = "2.21"
pydantic = "2.6.3"
pydantic-extra-types = "2.5.0"
pydantic-settings = "2.2.1"
pydantic-core = "2.16.3"
pyjwt = "2.8.0"
pyproject-hooks = "1.1.0"
python-dotenv = "1.0.1"
python-multipart = "0.0.7"
pywin32-ctypes = "0.2.2"
pyyaml = "6.0.1"
rapidfuzz = "3.9.3"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
sniffio = "1.3.1"
sqlalchemy = "2.0.27"
starlette = "0.36.3"
tomli = "2.0.1"
tomlkit = "0.12.5"
trove-classifiers = "2024.5.22"
typing-extensions = "4.10.0"
ujson = "5.9.0"
urllib3 = "2.2.2"
uvicorn = "0.27.1"
virtualenv = "20.26.3"
watchfiles = "0.21.0"
websockets = "12.0"
zipp = "3.19.2"
yookassa = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
